backend/requirements.txt
# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]

# Main API Dependencies
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart==0.0.6
pydantic>=2.10.0
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
httpx  # Used by OpenAI client in main.py

# PDF Generation Dependencies
# Note: pdflatex isn't a pip package, it should be installed via system package manager
jinja2>=2.11.2  # Template engine for LaTeX templating
python-dateutil>=2.8.2  # Date parsing and manipulation

# Automation
playwright>=1.41.0
beautifulsoup4>=4.12.0

# If you need NLP capabilities later, you can add: spacy>=3.0.0 and use python -m spacy download en_core_web_sm 

boto3==1.34.11

# Database Dependencies
sqlalchemy>=2.0.0
alembic>=1.13.0
apify-client